# perf(rust,python): avoid quadratic `exclude` behaviour when selecting against dtypes and/or wildcards