@mcrumiller 2023-05-20
What library did you use for the figure?