@alexander-beedie 2023-05-20
> What library did you use for the figure?

An elderly copy of Excel on the Mac 🤣